Add automated patient data and version control integration
Implemented a script for auto-generation of patient data in 'patient_names.txt', creating a more efficient way to populate the file for testing. Also added a new '.xml' file to enable Git as the version control system in the '.idea' directory, supporting better project tracking. Furthermore, a new python file 'Diseases.py' was added to store and manage diseases related data. The enhancements allow for a more organized workspace and smoother testing operations.